Enhance chat application: implement user management with join and disconnect events, improve message handling, and refactor client code for better structure
